Update BSFTest.java
change for error

+++ src/test/java/org/apache/bsf/BSFTest.java
@@ -140,7 +140,7 @@ public class BSFTest extends BSFEngineTestCase {
 
     public void testDeclareBean() {
         try {
-            bsfManager.declareBean("foo", new Integer(1), Integer.class);
+            bsfManager.declareBean("foo", new Integer(5), Integer.class);
         }
         catch (final Exception e) {
             fail(failMessage("declareBean() test failed", e));
